requirements.txt
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
django-rest-framework==0.1.0 ; python_version >= "3.11" and python_version < "4.0"
django==4.2.4 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.7 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
